+++ src/test/java/study/lotto/LottoTest.java
@@ -1,18 +1,22 @@
 package study.lotto;
 
+import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import study.lotto.model.*;
+import study.lotto.model.Lotto;
+import study.lotto.model.LottoList;
+import study.lotto.model.LottoNumber;
+import study.lotto.model.WinningLotto;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class LottoTest {
 
@@ -37,11 +41,12 @@ public class LottoTest {
     void generate_lotto() {
         Lotto lotto = Lotto.generate();
 
-        assertThat(lotto.getLottoNumbers())
+        SoftAssertions softly = new SoftAssertions();
+        softly.assertThat(lotto.getLottoNumbers())
                 .allMatch(lottoNumber -> lottoNumber.getNumber()>0 && lottoNumber.getNumber()<46);
-
-        assertThat(lotto.getLottoNumbers().size())
+        softly.assertThat(lotto.getLottoNumbers().size())
                 .isEqualTo(6);
+        softly.assertAll();
     }
 
     @DisplayName("당첨 번호 대조 테스트")
@@ -63,38 +68,10 @@ public class LottoTest {
     void create_lotto_list() {
         LottoList lottoList = LottoList.create(4);
 
-        assertThat(lottoList.getLottoList().size()).isEqualTo(4);
-
-        assertThat(lottoList.getLottoList())
+        SoftAssertions softly = new SoftAssertions();
+        softly.assertThat(lottoList.getLottoList().size()).isEqualTo(4);
+        softly.assertThat(lottoList.getLottoList())
                 .allMatch(Objects::nonNull);
-    }
-
-    @DisplayName("로또 당첨금 계산")
-    @ParameterizedTest
-    @ValueSource(strings = { "1, 2, 3, 4, 5, 6", "2, 3, 4, 7, 9, 12" })
-    void check_prize(String input) {
-        List<LottoNumber> lottoNumbers = Stream.of(1, 3, 5, 7, 9, 11).map(LottoNumber::new).collect(Collectors.toList());
-        Lotto lotto = new Lotto(lottoNumbers);
-
-        String[] split = input.split(", ");
-        WinningLotto winningLotto = WinningLotto.convertToWinningLotto(split);
-
-        LottoRank lottoRank = lotto.getPrize(winningLotto);
-
-        assertThat(lottoRank.getPrize())
-                .isEqualTo(5000);
-    }
-
-    @DisplayName("수익률 계산")
-    @Test
-    void calculate_earning_rate() {
-        LottoList lottoList = LottoList.create(10);
-
-        String[] split = "1, 2, 3, 4, 5, 6".split(", ");
-        WinningLotto winningLotto = WinningLotto.convertToWinningLotto(split);
-        Statistics statistics = Statistics.createStatistics(4000, lottoList, winningLotto);
-
-        assertThat(statistics.getEarningRate())
-                .isGreaterThanOrEqualTo(0);
+        softly.assertAll();
     }
 }
